refact: move routers to gateway package
