+++ insonmnia/gateway/router_nonlinux.go
@@ -1,6 +1,6 @@
 // +build !linux
 
-package hub
+package gateway
 
 import (
 	"context"
